+++ drift_dev/test/cli/migrate_test.dart
@@ -262,7 +262,8 @@ dependency_overrides:
   moor:
     path: /foo/bar
   moor_generator:
-    hosted: foo
+    hosted:
+      url: foo
     version: ^1.2.3
 ''');
 
@@ -288,7 +289,8 @@ dependency_overrides:
   drift:
     path: /foo/bar
   drift_dev:
-    hosted: foo
+    hosted:
+      url: foo
     version: ^1.2.3
 '''),
     ]).validate();